Replace RSS latest/image/succession commands with API
This makes thread-based games use the API to fetch the latest BYC post,
which need some BBCode parsing and optionally BYC game state image
generation in order to present a result. However, it is more stable than
RSS which does not seem to honor updating according to the latest post.