Filter search results with seed constraints
Locations, IC and Infiltration now hide results that are not in play
(unused boards and (non)errata'd cards) unless they are the only viable
result found (out of three for sanity). Only uses cached post seeds for
now since it would be too resourceful to download a new post when the
constraints refer to properties selected during setup.

Limited to locations and errata'd cards because those often cause most
confusion when unused versions show up. Could be used to filter other
cards later on, but care should be taken that this does not cause
completely irrelevant results to show up instead.

- Add "location" as alias for "board" in search queries
- Unbreak location attachment without bbox due to missing "image" field