Bot: Include allowed mentions in messages, discord-BGG user link
Upgrade to discord.py 1.4+ to make use of the allowed mentions field.
This contains all the mentions that we actually placed in a message, and
otherwise it is set to not allow those mentions. This makes it "safer"
to post messages from an external source which might have someone put
@everyone in it.

Additionally, improve username replacements when retrieving the latest
thread message so that a mapping between BGG and discord can be made.