Bot: Initial ping command
Pings roles mentioned in bold within the latest BGG post (other than the
post author's role) in an attempt to indicate what they need to do. This
also works fairly decent already with BYC skill check tokens, in that it
only pings the user up to play. Other situations are not considered yet.