Use banner image roles for ping mention exclusion
The banner image can be used to indicate the role of the user playing,
which could be used as a fallback in case we don't find the post author
using a config lookup. This helps with under-/misconfiguration, where
the post author could be pinged (for one bold message) when they should
not.

Added some proper sanity checks where images were used, initialize
variables in BBCode and avoid requesting text format for image often.