BYC: Properly handle modern dialog prompt styles
The code was assuming the options were always dialog inputs rather than
buttons, but fortunately if the input is missing it will attempt
a button press instead. This however fails in some styles where buttons
could not be found by nth-child selector, so instead track the dialog
button elements through the loop we already do for it. Also if we want
to press the last button, then ask the script to press the 0 - 1th
button instead of expecting there to be only one button.

This is only relevant for some BYC-on-BGG operations where we use the
script to extract a game state (.image, and .state in command-line)
because the BYC-in-bot operation does not allow users to permanently
change their prompt styles at the moment (if they somehow do it gets
reset whenever a public state or a new series of actions is made).