Ping: Mention current player
If we're in the Receive Skills or Movement phase, mention the current
player name (in favor of any bold text pings). Exception is the initial
draw in round one. This should ping any player whose turn is up, even if
they have a variable draw, and reduces the verbosity for those who
received their cards.