Add seed constraints for character cards
This uses an expression parser that evaluates an expression using seed
variables. Per-expansion constraints are combined with card-specific
constraints. Additionally, we track the alternate version of some
characters so that searches more often find the one that's in play.

Correct some Cylon Leader expansions and keep them separate from the
others, so that they have their own seed constraint.