Ping: Remove mentions of the player's title
(Only the title at the end of the post count, so losing of titles may
lead to mentioning it anyway). This should reduce pings for the current
user somewhat. This takes into account the seed titles or the banner
titles if we cannot look into the seed for some reason. We do not keep
track of discord roles here though, that could be an improvement as well
(but the logic to detect the roles is already quite complex).