Ping: Skip interrupts of skill check if it started
We detect if the interrupts belong to a skill check and see if the
skill check has started by checking if anything was played in it (like
destiny) or anyone has passed or played yet.

This follows BYC conventions of allowing players to play interrupts as
long as nobody has interacted with the skill check, even when a player
displays the skill check token when a Guts and Initiative was played as
an interrupt.