Flavor text and image characters, episode sources
This introduces data for parts of the cards the have no meaning for the
actual BGG game and are not in use for any game-related bot responses,
but may be relevant for extracting statistics/groupings from cards.
Nevertheless, there is no immediate reason for these additions,
notwithstanding that it's nice to have around for later use.