Allow asterisk in interrupt/skillcheck token topic
An asterisk may now appear in the topic of the token, as long as it is
not followed by one. This makes it possible to extract the topic of an
interrupt/skillcheck from a crisis that has a consequence icon on it.
(The crises with variations where the jump icon may appear do not have
crises or interrupts.)